requirements.txt
pyyaml
requests
requests-oauthlib
beautifulsoup4
Flask
apscheduler